Feature/s05 p31 a204 25 방 생성하기

+++ backend/northernplanet/src/main/java/com/pyhu/northernplanet/db/repository/ParticipantRepository.java
@@ -10,11 +10,11 @@ import org.springframework.stereotype.Repository;
 public interface ParticipantRepository extends JpaRepository<Participants, Integer> {
 
   @Transactional
-  List<Participants> findByrooms_roomId(int room_id);
+  List<Participants> findByrooms_roomId(Long room_id);
 
   @Transactional
-  List<Participants> findByusers_userId(int user_id);
+  List<Participants> findByusers_userId(Long user_id);
 
   @Transactional
-  int deleteAllByRooms_RoomId(int roomId);
+  int deleteAllByRooms_RoomId(Long roomId);
 }
